refactor: 도착위치까지의 경로를 구하는 기능 ChessBoardPosition에게 위임 - Bishop, Queen, Rook에서 중복되는 코드 제거
