+++ src/main/java/chess/domain/piece/Bishop.java
@@ -3,8 +3,6 @@ package chess.domain.piece;
 import chess.domain.ChessBoard;
 import chess.domain.ChessBoardPosition;
 import chess.domain.Team;
-import java.util.ArrayList;
-import java.util.List;
 
 public class Bishop extends ChessPiece {
     private static final double SCORE = 3.0;
@@ -28,29 +26,9 @@ public class Bishop extends ChessPiece {
 
     private boolean isUnobstructed(ChessBoardPosition sourcePosition, ChessBoardPosition targetPosition,
                                    ChessBoard board) {
-        return createPathPositions(sourcePosition, targetPosition)
+        return sourcePosition.createPathPositions(targetPosition)
                 .stream()
                 .noneMatch(board::existChessPieceAt);
-    }
-
-    private List<ChessBoardPosition> createPathPositions(ChessBoardPosition sourcePosition,
-                                                         ChessBoardPosition targetPosition) {
-        int rowUnitChange = calculateUnitChange(targetPosition.getRow(), sourcePosition.getRow());
-        int columnUnitChange = calculateUnitChange(targetPosition.getColumn(), sourcePosition.getColumn());
-        List<ChessBoardPosition> pathPositions = new ArrayList<>();
-        ChessBoardPosition currentBoardPosition = sourcePosition.move(columnUnitChange, rowUnitChange);
-        while (!currentBoardPosition.equals(targetPosition)) {
-            pathPositions.add(currentBoardPosition);
-            currentBoardPosition = currentBoardPosition.move(columnUnitChange, rowUnitChange);
-        }
-        return pathPositions;
-    }
-
-    private int calculateUnitChange(int source, int target) {
-        if (source == target) {
-            return 0;
-        }
-        return (source - target) / Math.abs(source - target);
     }
 
     private int calculateRowDistance(int highRow, int lowRow) {
